refactor: split startup metadata and notification owners
