+++ rustfs/src/lib.rs
@@ -68,11 +68,13 @@ pub mod profiling;
 pub mod protocols;
 pub mod runtime_capabilities;
 pub mod server;
+pub mod startup_bucket_metadata;
 pub(crate) mod startup_embedded;
 pub mod startup_entrypoint;
 pub mod startup_fs_guard;
 pub mod startup_iam;
 pub mod startup_lifecycle;
+pub mod startup_notification;
 pub mod startup_optional_runtime_sidecars;
 pub mod startup_optional_runtimes;
 pub mod startup_preflight;
